--- magicgui/widgets/__init__.py
@@ -1,38 +1,0 @@
-from ._widget_wrappers import CategoricalWidget, Widget
-from ._widgets import (
-    CheckBox,
-    ComboBox,
-    Container,
-    DateTimeEdit,
-    FileEdit,
-    FloatSlider,
-    FloatSpinBox,
-    Label,
-    LineEdit,
-    LogSlider,
-    PushButton,
-    RadioButton,
-    Slider,
-    SpinBox,
-    TextEdit,
-)
-
-__all__ = [
-    "CategoricalWidget",
-    "CheckBox",
-    "ComboBox",
-    "Container",
-    "DateTimeEdit",
-    "FileEdit",
-    "FloatSlider",
-    "FloatSpinBox",
-    "Label",
-    "LineEdit",
-    "LogSlider",
-    "PushButton",
-    "RadioButton",
-    "Slider",
-    "SpinBox",
-    "TextEdit",
-    "Widget",
-]
